refactor: 캐러셀 컴포넌트 파일명 변경
