+++ packages/design-system/src/components/index.ts
@@ -6,11 +6,11 @@ export { default as FloatingButton } from './floating-button/floating-button';
 export { default as ToastContainer } from './toast/toast-container';
 export { toast } from './toast/utils/toast';
 export { default as Header } from './header/header';
+export { default as TicketingCarousel } from './ticketing-carousel/ticketingCarousel-carousel';
 export { default as PerformanceCarousel } from './performance-carousel/performance-carousel';
-export { default as TopCarousel } from './top-carousel/top-carousel';
 export { default as ArtistCard } from './artist-card/artist-card';
 export { default as FestivalCard } from './festival-card/festival-card';
 export { default as Spacing } from './spacing/spacing';
 export { default as SearchBar } from './search-bar/search-bar';
-export { default as ProgressBar } from './performance-carousel/progress-bar/progress-bar';
-export { default as InfoBUtton } from './performance-carousel/info-button/info-button';
+export { default as ProgressBar } from './ticketing-carousel/progress-bar/progress-bar';
+export { default as InfoBUtton } from './ticketing-carousel/info-button/info-button';
