+++ packages/design-system/src/components/performance-carousel/dots.css.ts
