+++ packages/design-system/src/components/performance-carousel/performance-carousel.stories.tsx
@@ -1,29 +1,92 @@
-import { Meta } from '@storybook/react';
+import type { Meta } from '@storybook/react';
 import PerformanceCarousel from './performance-carousel';
-import { BANNER_DATA } from './mocks/bottom-banner-data';
 
-const meta: Meta<typeof PerformanceCarousel.Wrap> = {
+const MOCK_PERFORM_DATA = [
+  {
+    performanceId: 1,
+    type: 'concert',
+    title: '오아시스 내한공연',
+    subTitle: 'LIVE NATION PRESENTS COLDPLAY',
+    performanceAt: '2025.10.21',
+    posterUrl: 'https://dummyimage.com/197x262',
+  },
+  {
+    performanceId: 2,
+    type: 'festival',
+    title: '오아시스 내한공연',
+    subTitle: '',
+    performanceAt: '2025.10.21',
+    posterUrl: 'https://dummyimage.com/197x262',
+  },
+  {
+    performanceId: 3,
+    type: 'festival',
+    title: '오아시스 내한공연3',
+    subTitle: 'LIVE NATION PRESENTS COLDPLAY',
+    performanceAt: '2025.10.21',
+    posterUrl: 'https://dummyimage.com/197x262',
+  },
+  {
+    performanceId: 4,
+    type: 'festival',
+    title: '오아시스 내한공연4',
+    subTitle: '',
+    performanceAt: '2025.10.21',
+    posterUrl: 'https://dummyimage.com/197x262',
+  },
+  {
+    performanceId: 5,
+    type: 'festival',
+    title: '오아시스 내한공연5',
+    subTitle: 'LIVE NATION PRESENTS COLDPLAY',
+    performanceAt: '2025.10.21',
+    posterUrl: 'https://dummyimage.com/197x262',
+  },
+  {
+    performanceId: 6,
+    type: 'festival',
+    title: '오아시스 내한공연6',
+    subTitle: 'LIVE NATION PRESENTS COLDPLAY',
+    performanceAt: '2025.10.21',
+    posterUrl: 'https://dummyimage.com/197x262',
+  },
+  {
+    performanceId: 7,
+    type: 'festival',
+    title: '오아시스 내한공연7',
+    subTitle: '',
+    performanceAt: '2025.10.21',
+    posterUrl: 'https://dummyimage.com/197x262',
+  },
+];
+
+const meta: Meta<typeof PerformanceCarousel> = {
   title: 'Common/PerformanceCarousel',
-  component: PerformanceCarousel.Wrap,
+  component: PerformanceCarousel,
   parameters: {
     layout: 'centered',
   },
+  args: {
+    performData: MOCK_PERFORM_DATA,
+  },
   tags: ['autodocs'],
-  args: {},
 };
-
-const bannerData = BANNER_DATA?.data?.performances || [];
-const TotalIndexData = BANNER_DATA?.data?.performanceCount;
 
 export default meta;
 
-export const Default = () => {
-  return (
-    <div style={{ width: '375px' }}>
-      <PerformanceCarousel.Wrap
-        performances={bannerData}
-        indexData={TotalIndexData}
-      />
-    </div>
-  );
-};
+// decorators: [
+//   (Story) => {
+//     return (
+//       <div
+//         style={{
+//           width: '375px',
+//           height: '500px',
+//           background:
+//             'linear-gradient(180deg, #131433 -3.3%, #9747FF 89.71%)',
+//         }}
+//       >
+//         <Story />
+//       </div>
+//     );
+//   },
+// ],
