+++ packages/design-system/src/components/performance-carousel/slick.css.ts
