+++ packages/design-system/src/components/ticketing-carousel/hooks/use-carousel-data.ts
