+++ packages/design-system/src/components/ticketing-carousel/hooks/use-carousel-slide.ts
