+++ packages/design-system/src/components/ticketing-carousel/hooks/use-data-format.ts
