+++ packages/design-system/src/components/ticketing-carousel/info-button/info-button.css.ts
