+++ packages/design-system/src/components/ticketing-carousel/mocks/bottom-banner-data.ts
