+++ packages/design-system/src/components/ticketing-carousel/progress-bar/progress-bar.stories.tsx
