+++ packages/design-system/src/components/ticketing-carousel/progress-bar/progress-bar.tsx
